Split out fastar_util
